CRAYSAT-1314: Add product catalog query code
Add a new module, `query`, for clients to query the product catalog for
information about the installed product versions along with information
about the components provided by those products.

This module originally came from the `sat` Python package, but it was
moved here in order to allow it to be re-used across multiple
components. A few changes were made while migrating it here, including:

* Removed deprecated properties for getting the docker images provided
  by a product. The assumptions we were making for SAT do not broadly
  apply to other products, so these deprecated methods are likely to be
  confusing for any other product.
* Removed warning catching code that appears to no longer be necessary
  with the latest version of the `kubernetes` package.
* Make `product_version` argument of `ProductCatalog.get_product` method
  optional and have it default to getting the latest version.
* Add two properties to the `InstalledProductVersion` class, `active`,
  and `supports_active`.

Test Description:
Unit tests pass.

Installed into a Python virtualenv on baldar and ran a script which used
the `ProductCatalog` class to print information about installed versions
of products.

Modified a loftsman manifest that deploys `sat-cfs-install` to use this
version of the cray-product-catalog-update container image. Deployed the
manifest and ensured that the `cray-product-catalog` K8s ConfigMap was
updated.

Ran the container under `podman` to update the `cray-product-catalog`
K8s ConfigMap as is done in certain product installers, including SAT.

Used this code as a dependency in the `sat` python package and tested
`sat bootprep` and `sat showrev` commands that use it.