CASMCMS-7995: Fix the Artifactory upload paths
In v2 of the `jfrog` CLI, the default value for the `--flat` option
changed from true to false. If `--flat` is true, then files are uploaded
to the exact target path and their location in the source file system is
ignored. If `--flat` is false, then files are uploaded to a target that
includes their parent directories on the source file system.

This was resulting in Python modules and helm charts being uploaded into
an incorrect location. For example, Python module archives and wheels
were uploaded to a `dist` subdirectory in the pip repository and thus
could not be found by `pip`.

Fix this problem by specifying `--flat true` on `jfrog rt upload` calls
to keep the prior behavior.

Test Description:
Examined artifactory to see that helm charts and python modules get
published in the expected locations.  Checked that I could install the
new version of the python module with `pip`.