CRAYSAT-1314: Allow additional properties in product catalog entries
Allow additional properties to be present at the top level of the data
for a version of a product. Also allow additional properties at other
levels, including under the `component_versions` and `configuration`
properties and within elements under the `repositories` array.

This will provide flexibility for products to add additional properties
to their product catalog data while still allowing their other data
which conforms to the schema to be queried by the `query` module of
`cray_product_catalog`. This is helpful so that `sat showrev` can show
all products from the product catalog as long as they have a name and a
version.

The query code will not attempt to access any of these additional
properties, so they will not cause failures when querying those
products.

Since schema validation is only getting more permissive, this will not
cause any new failures when uploading product catalog data with
`catalog_update` when schema validation is enabled.

Test Description:
Tested this new version with `sat showrev`.