CASM-3589: Remove duplicate entries when merging dictionaries
This commit changes the behavior of the 'merge_dict' utility, such
that duplicate entries in a list are now replaced. Add a unit test
for this behavior.

The behavior was changed to prevent an infinite loop in the product
catalog update process. The main loop in update_catalog breaks out
of the loop only if merging in new content would do nothing (i.e.;
it does a merge of the input data into the current catalog data and
checks if it was unchanged. However, when merging a dictionary
containing lists that do not remove duplicates, the resulting data
will always change, leading to the process looping forever.

In addition to this issue in the logic, it probably makes more sense
to not allow duplicate lists of dictionaries in the catalog. For
example, if a partial install were to insert some of the repositories
for a product, we likely would not want to have duplicate entries. This
could possibly expose some edge cases involving installing multiple
of the same version of a product with changed component data, though
this is unlikely to happen in practice.

This commit adds a change log entry and declares a new version, 1.8.2.
As a part of updating the change log this commit changes a previous
version declaration (which was never tagged or released) to be 1.8.1
rather than 1.7.1.

Test Description:
* Unit test for the utility function.
* I added a unit test on the feature/add-unit-tests branch to test
  more thoroughly, but have not included it here.
* Using minikube, I tested inserting repository data into a local
  cray-product-catalog configmap without this change, reproducing
  the infinite loop, then with this change, showing it could be
  inserted.